+++ src/main/java/roomescape/controller/request/ReservationTimeWebRequest.java
@@ -1,4 +1,4 @@
-package roomescape.controller.dto;
+package roomescape.controller.request;
 
 public record ReservationTimeWebRequest(Long id, String startAt) {
