+++ src/main/java/roomescape/controller/request/ReservationWebRequest.java
@@ -1,4 +1,4 @@
-package roomescape.controller.dto;
+package roomescape.controller.request;
 
 public record ReservationWebRequest(String name, String date, Long timeId, Long themeId) {
