+++ src/main/java/roomescape/controller/request/ThemeWebRequest.java
@@ -1,4 +1,4 @@
-package roomescape.controller.dto;
+package roomescape.controller.request;
 
 public record ThemeWebRequest(String name, String description, String thumbnail) {
