+++ src/main/java/roomescape/service/ReservationService.java
@@ -12,7 +12,7 @@ import roomescape.domain.Theme;
 import roomescape.domain.ThemeRepository;
 import roomescape.exception.DuplicatedModelException;
 import roomescape.exception.PastReservationException;
-import roomescape.service.dto.ReservationAppRequest;
+import roomescape.service.request.ReservationAppRequest;
 
 @Service
 public class ReservationService {
