+++ src/main/java/roomescape/service/ReservationTimeService.java
@@ -10,8 +10,8 @@ import roomescape.domain.ReservationTime;
 import roomescape.domain.ReservationTimeRepository;
 import roomescape.exception.DuplicatedModelException;
 import roomescape.exception.ReservationExistsException;
-import roomescape.service.dto.ReservationTimeAppRequest;
-import roomescape.service.dto.ReservationTimeAppResponse;
+import roomescape.service.request.ReservationTimeAppRequest;
+import roomescape.service.response.ReservationTimeAppResponse;
 
 @Service
 public class ReservationTimeService {
