+++ src/main/java/roomescape/service/request/ReservationAppRequest.java
@@ -1,4 +1,4 @@
-package roomescape.service.dto;
+package roomescape.service.request;
 
 public record ReservationAppRequest(String name, String date, Long timeId, Long themeId) {
