+++ src/main/java/roomescape/service/request/ReservationTimeAppRequest.java
@@ -1,4 +1,4 @@
-package roomescape.service.dto;
+package roomescape.service.request;
 
 public record ReservationTimeAppRequest(String startAt) {
