+++ src/main/java/roomescape/service/response/ReservationTimeAppResponse.java
@@ -1,4 +1,4 @@
-package roomescape.service.dto;
+package roomescape.service.response;
 
 import com.fasterxml.jackson.annotation.JsonFormat;
 import java.time.LocalTime;
